Ignore vendor folder for lint and style

+++ .eslintrc.js
@@ -5,7 +5,7 @@ module.exports = {
     rules: {
         indent: ["error", 4],
         "linebreak-style": ["error", "unix"],
-        quotes: ["error", "single", {avoidEscape: true, allowTemplateLiterals: true}],
+        quotes: ["error", "double", {avoidEscape: true, allowTemplateLiterals: true}],
         "quote-props": ["error", "consistent"],
         semi: ["error", "always"],
     },
